starboard: Rename memory reuse allocators
Rename InPlaceReuseAllocatorBase to EmbeddedMetadataReuseAllocatorBase
and ReuseAllocatorBase to ExternalMetadataReuseAllocatorBase to clearly
reflect their metadata tracking behavior.

Bug: 454441375
