+++ starboard/common/BUILD.gn
@@ -51,18 +51,20 @@ static_library("common") {
     "device_type.h",
     "drm.cc",
     "drm.h",
+    "embedded_metadata_reuse_allocator_base.cc",
+    "embedded_metadata_reuse_allocator_base.h",
     "expected.h",
     "experimental/media_buffer_pool.cc",
     "experimental/media_buffer_pool.h",
     "experimental/media_buffer_pool_bidirectional_reuse_allocator.h",
     "experimental/media_buffer_pool_memory_allocator.h",
+    "external_metadata_reuse_allocator_base.cc",
+    "external_metadata_reuse_allocator_base.h",
     "file.cc",
     "file.h",
     "fixed_no_free_allocator.cc",
     "fixed_no_free_allocator.h",
     "flat_map.h",
-    "in_place_reuse_allocator_base.cc",
-    "in_place_reuse_allocator_base.h",
     "log.cc",
     "media.cc",
     "media.h",
@@ -82,8 +84,6 @@ static_library("common") {
     "ref_counted.cc",
     "ref_counted.h",
     "result.h",
-    "reuse_allocator_base.cc",
-    "reuse_allocator_base.h",
     "scoped_timer.cc",
     "scoped_timer.h",
     "semaphore.cc",
