+++ starboard/common/experimental/media_buffer_pool_bidirectional_reuse_allocator.h
@@ -21,7 +21,7 @@
 #include "starboard/common/bidirectional_fit_reuse_allocator.h"
 #include "starboard/common/experimental/media_buffer_pool.h"
 #include "starboard/common/experimental/media_buffer_pool_memory_allocator.h"
-#include "starboard/common/reuse_allocator_base.h"
+#include "starboard/common/external_metadata_reuse_allocator_base.h"
 
 namespace starboard {
 namespace experimental {
@@ -86,7 +86,7 @@ class MediaBufferPoolBidirectionalReuseAllocator : public Allocator {
 
  private:
   MediaBufferPoolMemoryAllocator fallback_allocator_;
-  BidirectionalFitReuseAllocator<ReuseAllocatorBase>
+  BidirectionalFitReuseAllocator<ExternalMetadataReuseAllocatorBase>
       bidirectional_fit_reuse_allocator_;
 };
