+++ starboard/common/external_metadata_reuse_allocator_base.cc
@@ -1,4 +1,4 @@
-// Copyright 2025 The Cobalt Authors. All Rights Reserved.
+// Copyright 2014 The Cobalt Authors. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "starboard/common/in_place_reuse_allocator_base.h"
+#include "starboard/common/external_metadata_reuse_allocator_base.h"
 
 #include <algorithm>
 #include <cstddef>
@@ -25,6 +25,10 @@
 namespace starboard {
 
 namespace {
+
+// Minimum block size to avoid extremely small blocks inside the block list and
+// to ensure that a zero sized allocation will return a non-zero sized block.
+const size_t kMinBlockSizeBytes = 16;
 
 int ceil_power_2(int i) {
   SB_DCHECK_GE(i, 0);
@@ -41,7 +45,8 @@ int ceil_power_2(int i) {
 
 }  // namespace
 
-bool InPlaceReuseAllocatorBase::MemoryBlock::Merge(const MemoryBlock& other) {
+bool ExternalMetadataReuseAllocatorBase::MemoryBlock::Merge(
+    const MemoryBlock& other) {
   SB_DCHECK_GE(fallback_allocation_index_, 0);
   SB_DCHECK_GE(other.fallback_allocation_index_, 0);
 
@@ -62,7 +67,7 @@ bool InPlaceReuseAllocatorBase::MemoryBlock::Merge(const MemoryBlock& other) {
   return false;
 }
 
-bool InPlaceReuseAllocatorBase::MemoryBlock::CanFulfill(
+bool ExternalMetadataReuseAllocatorBase::MemoryBlock::CanFulfill(
     size_t request_size,
     size_t alignment) const {
   SB_DCHECK_GE(fallback_allocation_index_, 0);
@@ -73,21 +78,16 @@ bool InPlaceReuseAllocatorBase::MemoryBlock::CanFulfill(
   return size_ >= aligned_size;
 }
 
-void InPlaceReuseAllocatorBase::MemoryBlock::Allocate(size_t request_size,
-                                                      size_t alignment,
-                                                      bool allocate_from_front,
-                                                      MemoryBlock* allocated,
-                                                      MemoryBlock* free) const {
+void ExternalMetadataReuseAllocatorBase::MemoryBlock::Allocate(
+    size_t request_size,
+    size_t alignment,
+    bool allocate_from_front,
+    MemoryBlock* allocated,
+    MemoryBlock* free) const {
   SB_DCHECK_GE(fallback_allocation_index_, 0);
   SB_DCHECK(allocated);
   SB_DCHECK(free);
   SB_DCHECK(CanFulfill(request_size, alignment));
-
-  // Minimum block size to avoid extremely small blocks inside the block list
-  // and to ensure that a zero sized allocation will return a non-zero sized
-  // block.
-  // TODO: b/399430536 - Cobalt: Tune this for a better value.
-  const size_t kMinBlockSizeBytes = 16 + sizeof(BlockMetadata);
 
   // First we assume that the block is just enough to fulfill the allocation and
   // leaves no free block.
@@ -134,23 +134,13 @@ void InPlaceReuseAllocatorBase::MemoryBlock::Allocate(size_t request_size,
   }
 }
 
-void* InPlaceReuseAllocatorBase::Allocate(size_t size) {
+void* ExternalMetadataReuseAllocatorBase::Allocate(size_t size) {
   return Allocate(size, 1);
 }
 
-void* InPlaceReuseAllocatorBase::Allocate(size_t size, size_t alignment) {
-  SB_DCHECK_EQ(sizeof(BlockMetadata) % alignment, 0U);
-
-  if (!pending_frees_.empty()) {
-    for (auto address_to_free : pending_frees_) {
-      bool freed = TryFree(address_to_free);
-      SB_DCHECK(freed);
-    }
-    pending_frees_.clear();
-  }
-
-  size = AlignUp(std::max(size, kMinAlignment), kMinAlignment) +
-         sizeof(BlockMetadata);
+void* ExternalMetadataReuseAllocatorBase::Allocate(size_t size,
+                                                   size_t alignment) {
+  size = AlignUp(std::max(size, kMinAlignment), kMinAlignment);
   alignment = AlignUp(std::max<size_t>(alignment, 1), kMinAlignment);
 
   bool allocate_from_front;
@@ -180,63 +170,18 @@ void* InPlaceReuseAllocatorBase::Allocate(size_t size, size_t alignment) {
     SB_DCHECK(free_block.address());
     AddFreeBlock(free_block);
   }
-
-  SB_DCHECK_EQ(
-      reinterpret_cast<intptr_t>(allocated_block.address()) % alignment, 0U);
-  SB_DCHECK_EQ(sizeof(BlockMetadata) % alignment, 0U);
-
-  void* user_address =
-      static_cast<uint8_t*>(allocated_block.address()) + sizeof(BlockMetadata);
-
-  AddAllocatedBlock(allocated_block);
+  void* user_address = AlignUp(allocated_block.address(), alignment);
+  AddAllocatedBlock(user_address, allocated_block);
 
   return user_address;
 }
 
-void InPlaceReuseAllocatorBase::Free(void* memory) {
-  if (memory == nullptr) {
-    return;
-  }
-
-  pending_frees_.push_back(memory);
-
-  if (pending_frees_.size() == total_allocated_blocks_) {
-    if (total_allocated_blocks_ > 32) {
-      SB_LOG_IF(INFO, total_allocated_blocks_ > 2048)
-          << "Batched free triggered for " << total_allocated_blocks_
-          << " blocks.";
-      allocated_block_head_ = nullptr;
-      free_blocks_.clear();
-      pending_frees_.clear();
-
-      total_allocated_in_bytes_ = 0;
-      total_allocated_blocks_ = 0;
-
-      for (int i = 0; i < static_cast<int>(fallback_allocations_.size()); ++i) {
-        AddFreeBlock(MemoryBlock(i, fallback_allocations_[i].address,
-                                 fallback_allocations_[i].size));
-      }
-
-      if (enable_decommit_on_idle_) {
-        SB_LOG(INFO) << "Batched free triggered idle state, decommitting "
-                     << fallback_allocations_.size()
-                     << " fallback allocations.";
-        for (const auto& fallback_allocation : fallback_allocations_) {
-          fallback_allocator_->Decommit(fallback_allocation.address,
-                                        fallback_allocation.size);
-        }
-      }
-    } else {
-      for (auto address_to_free : pending_frees_) {
-        bool freed = TryFree(address_to_free);
-        SB_DCHECK(freed);
-      }
-      pending_frees_.clear();
-    }
-  }
+void ExternalMetadataReuseAllocatorBase::Free(void* memory) {
+  bool result = TryFree(memory);
+  SB_DCHECK(result);
 }
 
-void InPlaceReuseAllocatorBase::PrintAllocations(
+void ExternalMetadataReuseAllocatorBase::PrintAllocations(
     bool align_allocated_size,
     int max_allocations_to_print) const {
   struct HistogramEntry {
@@ -252,10 +197,8 @@ void InPlaceReuseAllocatorBase::PrintAllocations(
   max_allocations_to_print = std::max(max_allocations_to_print, 1);
 
   // Logging the allocated blocks
-  BlockMetadata* head = allocated_block_head_;
-
-  while (head != nullptr) {
-    size_t size = head->size;
+  for (auto&& block : allocated_blocks_) {
+    size_t size = block.second.size();
     size_t size_as_key = align_allocated_size ? ceil_power_2(size) : size;
     HistogramEntry& entry = allocated_histogram[size_as_key];
 
@@ -263,32 +206,27 @@ void InPlaceReuseAllocatorBase::PrintAllocations(
     entry.min = std::min(entry.min, size);
     entry.max = std::max(entry.max, size);
     entry.total += size;
-
-    head = head->next;
   }
 
   int64_t allocated_percentage =
-      capacity_in_bytes_ == 0
-          ? 0
-          : static_cast<int64_t>(total_allocated_in_bytes_) * 100 /
-                capacity_in_bytes_;
-  SB_LOG(INFO) << "Allocated " << total_allocated_in_bytes_ << " bytes ("
+      capacity_ == 0 ? 0
+                     : static_cast<int64_t>(total_allocated_) * 100 / capacity_;
+  SB_LOG(INFO) << "Allocated " << total_allocated_ << " bytes ("
                << allocated_percentage << "%) from a pool of capacity "
-               << capacity_in_bytes_ << " bytes.  There are "
-               << capacity_in_bytes_ - total_allocated_in_bytes_
-               << " free bytes.";
-  SB_LOG(INFO) << "Total allocated block: " << total_allocated_blocks_;
+               << capacity_ << " bytes.  There are "
+               << capacity_ - total_allocated_ << " free bytes.";
+  SB_LOG(INFO) << "Total allocated block: " << allocated_blocks_.size();
 
   int lines = 0;
   size_t accumulated_blocks = 0;
 
   for (auto&& iter : allocated_histogram) {
     if (lines == max_allocations_to_print - 1 &&
-        static_cast<int>(allocated_histogram.size()) >
-            max_allocations_to_print) {
+        allocated_histogram.size() >
+            static_cast<size_t>(max_allocations_to_print)) {
       SB_LOG(INFO) << "\t[" << allocated_histogram.rbegin()->second.min << ", "
                    << iter.second.max
-                   << "] : " << total_allocated_blocks_ - accumulated_blocks;
+                   << "] : " << allocated_blocks_.size() - accumulated_blocks;
       break;
     }
 
@@ -319,7 +257,7 @@ void InPlaceReuseAllocatorBase::PrintAllocations(
 
   for (auto&& iter : free_histogram) {
     if (lines == max_allocations_to_print - 1 &&
-        static_cast<int>(free_histogram.size()) > max_allocations_to_print) {
+        free_histogram.size() > static_cast<size_t>(max_allocations_to_print)) {
       SB_LOG(INFO) << "\t[" << free_histogram.rbegin()->first << ", "
                    << iter.first
                    << "] : " << free_blocks_.size() - accumulated_blocks;
@@ -332,50 +270,26 @@ void InPlaceReuseAllocatorBase::PrintAllocations(
   }
 }
 
-bool InPlaceReuseAllocatorBase::TryFree(void* memory) {
+bool ExternalMetadataReuseAllocatorBase::TryFree(void* memory) {
   if (!memory) {
     return true;
   }
 
-  BlockMetadata* metadata = static_cast<BlockMetadata*>(memory) - 1;
-
-  if (metadata->signature != this || metadata->fallback_index < 0 ||
-      static_cast<size_t>(metadata->fallback_index) >=
-          fallback_allocations_.size()) {
+  AllocatedBlockMap::iterator it = allocated_blocks_.find(memory);
+  if (it == allocated_blocks_.end()) {
     return false;
   }
 
-  // Detach the node from the linked list
-  if (metadata->previous != nullptr) {
-    SB_DCHECK_EQ(metadata->previous->next, metadata);
-    metadata->previous->next = metadata->next;
-  }
-  if (metadata->next != nullptr) {
-    SB_DCHECK_EQ(metadata->next->previous, metadata);
-    metadata->next->previous = metadata->previous;
-  }
-
-  // Update the head, if the node is the head.
-  if (metadata->previous == nullptr) {
-    SB_DCHECK_EQ(metadata, allocated_block_head_);
-
-    allocated_block_head_ = metadata->next;
-
-    if (metadata->next) {
-      // It should have been updated during detaching above.
-      SB_DCHECK_EQ(metadata->next->previous, nullptr);
-    }
-  }
-
-  // Mark this block as free and update the allocated data accordingly.
-  MemoryBlock block(metadata->fallback_index, metadata, metadata->size);
+  // Mark this block as free and remove it from the allocated set.
+  const MemoryBlock& block = (*it).second;
   AddFreeBlock(block);
 
-  SB_DCHECK_LE(block.size(), total_allocated_in_bytes_);
-  total_allocated_in_bytes_ -= block.size();
-  --total_allocated_blocks_;
+  SB_DCHECK_LE(block.size(), total_allocated_);
+  total_allocated_ -= block.size();
 
-  if (enable_decommit_on_idle_ && total_allocated_in_bytes_ == 0) {
+  allocated_blocks_.erase(it);
+
+  if (enable_decommit_on_idle_ && total_allocated_ == 0) {
     SB_LOG(INFO) << "Allocator reached idle state, decommitting "
                  << fallback_allocations_.size() << " fallback allocations.";
     for (const auto& fallback_allocation : fallback_allocations_) {
@@ -387,7 +301,7 @@ bool InPlaceReuseAllocatorBase::TryFree(void* memory) {
   return true;
 }
 
-InPlaceReuseAllocatorBase::InPlaceReuseAllocatorBase(
+ExternalMetadataReuseAllocatorBase::ExternalMetadataReuseAllocatorBase(
     Allocator* fallback_allocator,
     size_t initial_capacity,
     size_t allocation_increment,
@@ -395,17 +309,17 @@ InPlaceReuseAllocatorBase::InPlaceReuseAllocatorBase(
     bool enable_decommit_on_idle)
     : fallback_allocator_(fallback_allocator),
       allocation_increment_(allocation_increment),
-      max_capacity_in_bytes_(max_capacity),
-      enable_decommit_on_idle_(enable_decommit_on_idle) {
+      max_capacity_(max_capacity),
+      enable_decommit_on_idle_(enable_decommit_on_idle),
+      capacity_(0),
+      total_allocated_(0) {
   if (initial_capacity > 0) {
     FreeBlockSet::iterator iter = ExpandToFit(initial_capacity, kMinAlignment);
     SB_DCHECK(iter != free_blocks_.end());
   }
-
-  pending_frees_.reserve(16 * 1024);
 }
 
-InPlaceReuseAllocatorBase::~InPlaceReuseAllocatorBase() {
+ExternalMetadataReuseAllocatorBase::~ExternalMetadataReuseAllocatorBase() {
   if (ExtraLogLevel() >= 1) {
     SB_LOG(INFO) << "Destroying reuse allocator ...";
     PrintAllocations(true, 16);
@@ -413,16 +327,16 @@ InPlaceReuseAllocatorBase::~InPlaceReuseAllocatorBase() {
 
   // Check that everything was freed.  Note that in some unit tests this may not
   // be the case.
-  SB_LOG_IF(ERROR, allocated_block_head_ != nullptr)
-      << total_allocated_blocks_ << " blocks still allocated.";
+  SB_LOG_IF(ERROR, allocated_blocks_.size() != 0)
+      << allocated_blocks_.size() << " blocks still allocated.";
 
   for (const auto& fallback_allocation : fallback_allocations_) {
     fallback_allocator_->Free(fallback_allocation.address);
   }
 }
 
-InPlaceReuseAllocatorBase::FreeBlockSet::iterator
-InPlaceReuseAllocatorBase::ExpandToFit(size_t size, size_t alignment) {
+ExternalMetadataReuseAllocatorBase::FreeBlockSet::iterator
+ExternalMetadataReuseAllocatorBase::ExpandToFit(size_t size, size_t alignment) {
   if (ExtraLogLevel() >= 1) {
     int capacity = GetCapacity();
     int allocated = GetAllocated();
@@ -443,8 +357,7 @@ InPlaceReuseAllocatorBase::ExpandToFit(size_t size, size_t alignment) {
   // fragmentation.
   if (allocation_increment_ > size) {
     size_to_try = allocation_increment_;
-    if (!max_capacity_in_bytes_ ||
-        capacity_in_bytes_ + size_to_try <= max_capacity_in_bytes_) {
+    if (!max_capacity_ || capacity_ + size_to_try <= max_capacity_) {
       ptr = fallback_allocator_->AllocateForAlignment(&size_to_try, alignment);
     }
   }
@@ -453,14 +366,13 @@ InPlaceReuseAllocatorBase::ExpandToFit(size_t size, size_t alignment) {
   // |size| instead for both cases.
   if (ptr == NULL) {
     size_to_try = size;
-    if (!max_capacity_in_bytes_ ||
-        capacity_in_bytes_ + size_to_try <= max_capacity_in_bytes_) {
+    if (!max_capacity_ || capacity_ + size_to_try <= max_capacity_) {
       ptr = fallback_allocator_->AllocateForAlignment(&size_to_try, alignment);
     }
   }
   if (ptr != NULL) {
-    fallback_allocations_.emplace_back(ptr, size_to_try);
-    capacity_in_bytes_ += size_to_try;
+    fallback_allocations_.push_back({ptr, size_to_try});
+    capacity_ += size_to_try;
     auto free_block_iter = AddFreeBlock(MemoryBlock(
         static_cast<int>(fallback_allocations_.size() - 1), ptr, size_to_try));
 
@@ -520,8 +432,7 @@ InPlaceReuseAllocatorBase::ExpandToFit(size_t size, size_t alignment) {
   //                     |           |
   // |free_address + free_size|  |aligned_address|
   size_t size_to_allocate = aligned_address + size - free_address - free_size;
-  if (max_capacity_in_bytes_ &&
-      capacity_in_bytes_ + size_to_allocate > max_capacity_in_bytes_) {
+  if (max_capacity_ && capacity_ + size_to_allocate > max_capacity_) {
     SB_LOG_IF(INFO, ExtraLogLevel() >= 1) << "Failed to expand.";
     return free_blocks_.end();
   }
@@ -531,8 +442,8 @@ InPlaceReuseAllocatorBase::ExpandToFit(size_t size, size_t alignment) {
     return free_blocks_.end();
   }
 
-  fallback_allocations_.emplace_back(ptr, size_to_allocate);
-  capacity_in_bytes_ += size_to_allocate;
+  fallback_allocations_.push_back({ptr, size_to_allocate});
+  capacity_ += size_to_allocate;
   AddFreeBlock(MemoryBlock(static_cast<int>(fallback_allocations_.size() - 1),
                            ptr, size_to_allocate));
   FreeBlockSet::iterator iter = free_blocks_.end();
@@ -563,33 +474,16 @@ InPlaceReuseAllocatorBase::ExpandToFit(size_t size, size_t alignment) {
   }
 }
 
-void InPlaceReuseAllocatorBase::AddAllocatedBlock(const MemoryBlock& block) {
-  BlockMetadata* metadata = static_cast<BlockMetadata*>(block.address());
-
-  metadata->signature = this;
-  metadata->fallback_index = block.fallback_allocation_index();
-  metadata->size = block.size();
-
-  // The node will be the new head, so let the existing head be its next node.
-  if (allocated_block_head_) {
-    SB_DCHECK_EQ(allocated_block_head_->previous, nullptr);
-
-    allocated_block_head_->previous = metadata;
-    metadata->next = allocated_block_head_;
-  } else {
-    metadata->next = nullptr;
-  }
-
-  // Now set it to the new head
-  metadata->previous = nullptr;
-  allocated_block_head_ = metadata;
-
-  total_allocated_in_bytes_ += block.size();
-  ++total_allocated_blocks_;
+void ExternalMetadataReuseAllocatorBase::AddAllocatedBlock(
+    void* address,
+    const MemoryBlock& block) {
+  SB_DCHECK(allocated_blocks_.find(address) == allocated_blocks_.end());
+  allocated_blocks_[address] = block;
+  total_allocated_ += block.size();
 }
 
-InPlaceReuseAllocatorBase::FreeBlockSet::iterator
-InPlaceReuseAllocatorBase::AddFreeBlock(MemoryBlock block_to_add) {
+ExternalMetadataReuseAllocatorBase::FreeBlockSet::iterator
+ExternalMetadataReuseAllocatorBase::AddFreeBlock(MemoryBlock block_to_add) {
   if (free_blocks_.size() == 0) {
     return free_blocks_.insert(block_to_add).first;
   }
@@ -600,7 +494,6 @@ InPlaceReuseAllocatorBase::AddFreeBlock(MemoryBlock block_to_add) {
   // if one exists.
   FreeBlockSet::iterator right_to_erase = free_blocks_.end();
   FreeBlockSet::iterator left_to_erase = free_blocks_.end();
-  FreeBlockSet::iterator insert_hint = it;
 
   if (it != free_blocks_.end()) {
     MemoryBlock right_block = *it;
@@ -620,18 +513,17 @@ InPlaceReuseAllocatorBase::AddFreeBlock(MemoryBlock block_to_add) {
   }
 
   if (right_to_erase != free_blocks_.end()) {
-    insert_hint = right_to_erase;
-    ++insert_hint;
     free_blocks_.erase(right_to_erase);
   }
   if (left_to_erase != free_blocks_.end()) {
     free_blocks_.erase(left_to_erase);
   }
 
-  return free_blocks_.insert(insert_hint, block_to_add);
+  return free_blocks_.insert(block_to_add).first;
 }
 
-void InPlaceReuseAllocatorBase::RemoveFreeBlock(FreeBlockSet::iterator it) {
+void ExternalMetadataReuseAllocatorBase::RemoveFreeBlock(
+    FreeBlockSet::iterator it) {
   free_blocks_.erase(it);
 }
